% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissipate.R
\name{dissipate}
\alias{dissipate}
\title{Calculates daily precipitation from accumulative precipitation}
\usage{
dissipate(data)
}
\arguments{
\item{data}{A vector of accumulative precipitation data}
}
\value{
A vector of daily precipitation
}
\description{
Calculates daily precipitation from accumulative precipitation
}

