% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelflow.R
\name{modelflow}
\alias{modelflow}
\title{Calculates a modeled streamflow}
\usage{
modelflow(data, meltcoefficient, runoffcoefficient, area, baseflow)
}
\arguments{
\item{data}{A data frame with a shortwave radiation, albedo, average temperature, percent snowcover, and precipitation}

\item{meltcoefficient}{(melt factor*snow runoff coefficient*radiation coefficient)}

\item{runoffcoefficient}{(between 0-1)}

\item{area}{Area of watershed in Km2}

\item{baseflow}{Discharge in cms}
}
\value{
vector of modeled flow
}
\description{
Calculates a modeled streamflow
}

